add baseItemModel, reduce code
