+++ napari/_qt/containers/tree/qt_tree_model.py
@@ -2,18 +2,17 @@ import logging
 import pickle
 from typing import Generic, List, Optional, Tuple, TypeVar, Union
 
-from qtpy.QtCore import QAbstractItemModel, QMimeData, QModelIndex, Qt
-from qtpy.QtWidgets import QWidget
+from qtpy.QtCore import QMimeData, QModelIndex, Qt
 
-from ....utils.events import disconnect_events
 from ....utils.tree import Group, Node
+from .._base_item_model import _BaseItemModel
 
 logger = logging.getLogger(__name__)
 NodeType = TypeVar("NodeType", bound=Node)
 NodeMIMEType = "application/x-tree-node"
 
 
-class QtNodeTreeModel(QAbstractItemModel, Generic[NodeType]):
+class QtNodeTreeModel(_BaseItemModel, Generic[NodeType]):
     """A concrete QItemModel for a tree of ``Node`` and ``Group`` objects.
 
     Key concepts and references:
@@ -26,12 +25,11 @@ class QtNodeTreeModel(QAbstractItemModel, Generic[NodeType]):
           <https://doc.qt.io/qt-5/qtwidgets-itemviews-simpletreemodel-example.html>`_
     """
 
-    def __init__(self, root: Group[NodeType], parent: QWidget = None):
-        super().__init__(parent)
-        self._root: Group[NodeType]
-        self.setRoot(root)
-
     # ########## Reimplemented Public Qt Functions ##################
+    def setRoot(self, root: Group[NodeType]):
+        if not isinstance(root, Group):
+            raise TypeError(f"root node must be an instance of {Group}")
+        super().setRoot(root)
 
     def canDropMimeData(
         self,
@@ -56,13 +54,6 @@ class QtNodeTreeModel(QAbstractItemModel, Generic[NodeType]):
             Whether we can accept a drop of data at ``row``, ``column``, ``parent``
         """
         return self.getItem(parent).is_group()
-
-    def columnCount(self, parent: QModelIndex) -> int:
-        """Returns the number of columns for the children of the given ``parent``.
-
-        In a tree, the number of columns is always 1.
-        """
-        return 1
 
     def data(self, index: QModelIndex, role: Qt.ItemDataRole):
         """Return data stored under ``role`` for the item at ``index``.
@@ -131,34 +122,12 @@ class QtNodeTreeModel(QAbstractItemModel, Generic[NodeType]):
             # for root
             return Qt.ItemIsDropEnabled
 
-        base_flags = (
-            Qt.ItemIsSelectable
-            | Qt.ItemIsEditable
-            | Qt.ItemIsUserCheckable
-            | Qt.ItemIsDragEnabled
-            | Qt.ItemIsEnabled
-            | Qt.ItemIsDropEnabled
-        )
+        base_flags = super().flags(index)
         if self.getItem(index).is_group():
-            return base_flags
-        return base_flags | Qt.ItemNeverHasChildren
-
-    def index(
-        self, row: int, column: int = 0, parent: QModelIndex = QModelIndex()
-    ) -> QModelIndex:
-        """Return index of the item specified by ``row``, ``column`` and ``parent`` index.
-
-        From Qt Docs: "When reimplementing this function in a subclass, call
-        ``createIndex()`` to generate model indexes that other components can use to refer
-        to items in your model.
-        """
-        if not self.hasIndex(row, column, parent):
-            return QModelIndex()
-
-        parentItem = self.getItem(parent)
-        if parentItem.is_group():
-            return self.createIndex(row, column, parentItem[row])  # type: ignore
-        return QModelIndex()
+            return (
+                base_flags | Qt.ItemIsDropEnabled
+            ) & ~Qt.ItemNeverHasChildren
+        return base_flags
 
     def mimeData(self, indices: List[QModelIndex]) -> Optional['NodeMimeData']:
         """Return an object containing serialized data corresponding to specified indexes.
@@ -195,7 +164,7 @@ class QtNodeTreeModel(QAbstractItemModel, Generic[NodeType]):
         return [NodeMIMEType, "text/plain"]
 
     def parent(self, index: QModelIndex) -> QModelIndex:
-        """Returns the parent of the model item with the given ``inde``x.
+        """Returns the parent of the model item with the given ``index``.
 
         If the item has no parent, an invalid QModelIndex is returned.
         """
@@ -211,114 +180,7 @@ class QtNodeTreeModel(QAbstractItemModel, Generic[NodeType]):
         row = parentItem.index_in_parent() or 0
         return self.createIndex(row, 0, parentItem)
 
-    def rowCount(self, parent: QModelIndex) -> int:
-        """Returns the number of rows under the given parent.
-
-        When the parent is valid it means that rowCount is returning the number of
-        children of parent.
-        """
-        try:
-            return len(self.getItem(parent))  # type: ignore
-        except TypeError:
-            return 0
-
-    def supportedDropActions(self) -> Qt.DropActions:
-        """Returns the drop actions supported by this model.
-
-        The default implementation returns Qt::CopyAction. We re-implement to support only
-        MoveAction. See also dropMimeData(), which must handles each operation.
-        """
-        return Qt.MoveAction
-
     # ########## New methods added for Group Model ##################
-
-    def setRoot(self, root: Group[NodeType]):
-        if not isinstance(root, Group):
-            raise TypeError(
-                "root node must be an instance of napari.utils.tree.Group"
-            )
-        current_root: Optional[Group[NodeType]] = getattr(self, "_root", None)
-        if root is current_root:
-            return
-        elif current_root is not None:
-            disconnect_events(self._root.events, self)
-
-        self._root = root
-        self._root.events.removing.connect(self._on_begin_removing)
-        self._root.events.removed.connect(self._on_end_remove)
-        self._root.events.inserting.connect(self._on_begin_inserting)
-        self._root.events.inserted.connect(self._on_end_insert)
-        self._root.events.moving.connect(self._on_begin_moving)
-        self._root.events.moved.connect(self._on_end_move)
-        self._root.events.connect(self._process_event)
-
-    def _process_event(self, event):
-        # for subclasses to handle NodeType-specific data
-        pass
-
-    def _on_begin_removing(self, event):
-        """Begins a row removal operation.
-
-        See Qt documentation: https://doc.qt.io/qt-5/qabstractitemmodel.html#beginRemoveRows
-        """
-        par, idx = self._split_nested_index(event.index)
-        self.beginRemoveRows(par, idx, idx)
-
-    def _on_end_remove(self, e):
-        self.endRemoveRows()
-
-    def _on_begin_inserting(self, event):
-        """Begins a row insertion operation.
-
-        See Qt documentation: https://doc.qt.io/qt-5/qabstractitemmodel.html#beginInsertRows
-        """
-        par, idx = self._split_nested_index(event.index)
-        self.beginInsertRows(par, idx, idx)
-
-    def _on_end_insert(self, e):
-        self.endInsertRows()
-
-    def _on_begin_moving(self, event):
-        """Begins a row move operation.
-
-        See Qt documentation: https://doc.qt.io/qt-5/qabstractitemmodel.html#beginMoveRows
-        """
-        src_par, src_idx = self._split_nested_index(event.index)
-        dest_par, dest_idx = self._split_nested_index(event.new_index)
-
-        logger.debug(
-            f"beginMoveRows({self.getItem(src_par)._node_name()}, {src_idx}, "
-            f"{self.getItem(dest_par)._node_name()}, {dest_idx})"
-        )
-
-        self.beginMoveRows(src_par, src_idx, src_idx, dest_par, dest_idx)
-
-    def _on_end_move(self, e):
-        self.endMoveRows()
-
-    def getItem(self, index: QModelIndex) -> NodeType:
-        """Return ``Node`` object for a given `QModelIndex`.
-
-        A null or invalid ``QModelIndex`` will return the root Node.
-        """
-        if index.isValid():
-            item = index.internalPointer()
-            if item is not None:
-                return item
-        return self._root
-
-    def findIndex(self, obj: NodeType) -> QModelIndex:
-        """Find the QModelIndex for a given ``Node`` object in the model."""
-        hits = self.match(
-            self.index(0),
-            Qt.UserRole,
-            obj,
-            1,
-            Qt.MatchExactly | Qt.MatchRecursive,
-        )
-        if hits:
-            return hits[0]
-        return QModelIndex()
 
     def nestedIndex(self, nested_index: Tuple[int, ...]) -> QModelIndex:
         """Return a QModelIndex for a given ``nested_index``."""
@@ -335,18 +197,6 @@ class QtNodeTreeModel(QAbstractItemModel, Generic[NodeType]):
             raise TypeError("nested_index must be an int or tuple of int.")
         return self.index(child, 0, parent)
 
-    def _split_nested_index(
-        self, nested_index: Union[int, Tuple[int, ...]]
-    ) -> Tuple[QModelIndex, int]:
-        """Given a nested index, return (nested_parent_index, row)."""
-        if isinstance(nested_index, int):
-            return QModelIndex(), nested_index
-        par = QModelIndex()
-        *_p, idx = nested_index
-        for i in _p:
-            par = self.index(i, 0, par)
-        return par, idx
-
     def hasGroups(self) -> bool:
         """Return true if the root has any Groups."""
         return any(node.is_group() for node in self._root)
@@ -358,7 +208,7 @@ class NodeMimeData(QMimeData):
         self.nodes: List[NodeType] = nodes or []
         if nodes:
             self.setData(NodeMIMEType, pickle.dumps(self.node_indices()))
-            self.setText(" ".join([node._node_name() for node in nodes]))
+            self.setText(" ".join(node._node_name() for node in nodes))
 
     def formats(self) -> List[str]:
         return [NodeMIMEType, "text/plain"]
